Using literal KaTeX HTML in all posts.
Previously, the KaTeX `renderToString` output could potentially be
mangled by Markdown processing (inline HTML is expected to do this, so
it's not a feature note a bug of Markdown). To work around this I
made a Markdown extension and hex-encoded the actual HTML in elements
of the form `<html-literal>{hex(_block}</html-literal>`.

This change is quite minimal. The only changes are

- The addition of `\\` character that had been removed, e.g. `\\\\``
  (a LaTeX newline) being collapsed to `\\`. This occurred in:
  - 2011/07/continued-fraction-expansions-of.html
  - 2011/08/conways-topograph-part-1.html
  - 2011/08/conways-topograph-part-2.html
  - 2011/08/conways-topograph-part-3.html
  - 2011/08/finding-fibonacci-golden-nuggets-part-2.html
  - 2011/08/finding-fibonacci-golden-nuggets.html
  - 2012/05/reverse-calculating-interest-rate.html
  - 2013/09/calculating-greatest-common-divisor.html
  - 2013/11/trigonometry-and-nested-radicals.html
- The addition of a missing `_{p^2}`. This was in part due to my
  old HTML de-mangling script.
  - 2013/08/some-fibonacci-fun-with-primes.html
- The removal of an anchor caused by a misinterpreted `\sqrt[3]` in
  `2016/11/newtons-failure.html`

Via commit a42285447966ece16d7cad8801c98269a0e92465 in
`bossylobster-blog`.